[TASK] logger-interface renamed just logger
